requirements.txt
python-dotenv>=0.19.0
openai>=1.0.0
# Add any other dependencies your project uses 